Updated CVR and PhoneNumber integers to strings Updated HourlyCost Integer to double

+++ PrisPilot/Models/Customer.cs
@@ -16,7 +16,7 @@ namespace PrisPilot.Models
         // We're also going to add it to our database as a bytearray anyways
         public byte[]? Logo { get; set; } = null;
         public string ContactPerson { get; set; } = string.Empty;
-        public int HourlyCost { get; set; } = 0;
+        public double HourlyCost { get; set; } = 0;
         
         public Customer()
         {
